Update authentication, models, and views across services
Refactored authentication with custom JWT settings and token serializers, extended field definitions in models, and improved API responses for better clarity. Cart operations now include accurate total calculations and item management, while product and customer services enhance permission handling and validation.

+++ product-service/product/views.py
@@ -1,11 +1,14 @@
 import logging
 import re
+import uuid
 
 import cloudinary.uploader
 from django.conf import settings
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from rest_framework_simplejwt.authentication import JWTTokenUserAuthentication
 
 from .models import ItemModel
 from .serializers import BookSerializer, SmartphoneSerializer, FashionSerializer, BaseItemSerializer
@@ -13,7 +16,7 @@ from .serializers import BookSerializer, SmartphoneSerializer, FashionSerializer
 cloudinary.config(
     cloud_name=settings.CLOUDINARY_CLOUD_NAME,
     api_key=settings.CLOUDINARY_API_KEY,
-    api_secret=settings.CLOUDINARY_API_SECRET,  # Click 'View API Keys' above to copy your API secret
+    api_secret=settings.CLOUDINARY_API_SECRET,
     secure=True
 )
 
@@ -24,38 +27,61 @@ def get_serializer_by_category(category):
         return BookSerializer
     elif category == "mobiles":
         return SmartphoneSerializer
-    elif category in ["fashions"]:
+    elif category == "fashions":
         return FashionSerializer
     return BaseItemSerializer
 
 
-# Create your views here.
 class AddItemView(APIView):
+    authentication_classes = [JWTTokenUserAuthentication]
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         try:
+            user = request.user
+
+            if not user or (user.role != "seller" and not user.is_superuser):
+                return Response({"success": False, "message": "Invalid seller"}, status=status.HTTP_400_BAD_REQUEST)
+
             data = request.data
             category = data.get("category", "").lower()
             if not category:
-                return Response({"error": "Invalid category"}, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"success": False, "message": "Invalid category"}, status=status.HTTP_400_BAD_REQUEST)
+
             serializer_class = get_serializer_by_category(category)
             serializer = serializer_class(data=data)
-
             if not serializer.is_valid():
-                return Response({"error": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"success": False, "errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
             image_urls = []
             if "images" in request.FILES:
                 uploaded_images = request.FILES.getlist("images")
                 for image in uploaded_images:
-                    upload_result = cloudinary.uploader.upload(image, folder=f"assets/{category}")
-                    image_urls.append(upload_result["secure_url"])
+                    try:
+                        upload_result = cloudinary.uploader.upload(
+                            image,
+                            public_id=f"{category}/{uuid.uuid4()}",
+                            folder=f"{category}",
+                            overwrite=True,
+                            transfromation=[
+                                {"width": 800, "height": 800, "crop": "limit", "quality": "auto"}
+                            ]
+                        )
+                        image_urls.append(upload_result["secure_url"])
+                    except Exception as e:
+                        return Response({"success": False, "message": "Error uploading image", "error": str(e)},
+                                        status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
+            logging.warning(user)
+            serializer.validated_data["seller_id"] = str(user.id)
             serializer.validated_data["image_urls"] = image_urls
             item_id = ItemModel.create_item(serializer.validated_data)
 
-            return Response({"item_id": item_id}, status=status.HTTP_201_CREATED)
+            return Response({"success": True, "message": "Item added successfully.", "data": {"item_id": item_id}},
+                            status=status.HTTP_201_CREATED)
         except Exception as e:
-            return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"success": False, "message": "Error adding item", "error": str(e)},
+                            status=status.HTTP_400_BAD_REQUEST)
 
 
 class ItemListView(APIView):
@@ -63,22 +89,50 @@ class ItemListView(APIView):
         items = ItemModel.get_all_items()
         for item in items:
             item["_id"] = str(item["_id"])
-        return Response({"items": items}, status=status.HTTP_200_OK)
+        return Response({"success": True, "message": "Items retrieved successfully.", "data": {"items": items}},
+                        status=status.HTTP_200_OK)
+
+
+class SellerItemListView(APIView):
+    authentication_classes = [JWTTokenUserAuthentication]
+    permission_classes = [IsAuthenticated]
+
+    def get(self, request):
+        seller_id = request.user.id
+        items = ItemModel.get_seller_items(str(seller_id))
+        for item in items:
+            item["_id"] = str(item["_id"])
+        return Response({"success": True, "message": "Items retrieved successfully.", "data": {"items": items}},
+                        status=status.HTTP_200_OK)
 
 
 class ItemDetailView(APIView):
+
+    def get_permissions(self):
+        if self.request.method == "GET":
+            return [AllowAny()]
+        else:
+            return [IsAuthenticated()] and [JWTTokenUserAuthentication()]
+
     def get(self, request, item_id):
-        logging.warning(item_id)
         item = ItemModel.get_item(item_id)
-        logging.warning(item)
         if item:
             item["_id"] = str(item["_id"])
-            return Response(item, status=status.HTTP_200_OK)
-        return Response({"error": "Item not found"}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"success": True, "message": "Item retrieved successfully.", "data": item},
+                            status=status.HTTP_200_OK)
+        return Response({"success": False, "message": "Item not found"}, status=status.HTTP_404_NOT_FOUND)
 
     def put(self, request, item_id):
-        data = request.data
+        user = request.user
+        item = ItemModel.get_item(item_id)
 
+        if not item:
+            return Response({"success": False, "message": "Item not found"}, status=status.HTTP_404_NOT_FOUND)
+
+        if str(item.get("seller_id")) != str(user.id) and not user.is_superuser:
+            return Response({"success": False, "message": "Permission denied."}, status=status.HTTP_403_FORBIDDEN)
+
+        data = request.data
         if "images" in request.FILES:
             image_urls = []
             uploaded_images = request.FILES.getlist("images")
@@ -88,21 +142,27 @@ class ItemDetailView(APIView):
             data["image_urls"] = image_urls
 
         ItemModel.update_item(item_id, data)
-        return Response({"item_id": item_id}, status=status.HTTP_200_OK)
+        return Response({"success": True, "message": "Item updated successfully.", "data": {"item_id": item_id}},
+                        status=status.HTTP_200_OK)
 
     def delete(self, request, item_id):
-        logging.warning(item_id)
+        user = request.user
         item = ItemModel.get_item(item_id)
-        logging.warning(item)
+
+        if not item:
+            return Response({"success": False, "message": "Item not found"}, status=status.HTTP_404_NOT_FOUND)
+
+        if str(item.get("seller_id")) != str(user.id) and not user.is_superuser:
+            return Response({"success": False, "message": "Permission denied."}, status=status.HTTP_403_FORBIDDEN)
+
         for image_url in item["image_urls"]:
-            logging.warning(image_url)
             match = re.search(r"/v\d+/(.+)\.\w+$", image_url)
             if match:
                 public_id = match.group(1)
                 cloudinary.uploader.destroy(public_id)
-                logging.warning(public_id)
+
         ItemModel.delete_item(item_id)
-        return Response({"success": "true"}, status=status.HTTP_200_OK)
+        return Response({"success": True, "message": "Item deleted successfully."}, status=status.HTTP_200_OK)
 
 
 class SearchItemView(APIView):
@@ -111,7 +171,10 @@ class SearchItemView(APIView):
         items = ItemModel.search_items({"name": {"$regex": query, "$options": "i"}})
         for item in items:
             item["_id"] = str(item["_id"])
-        return Response({"items": items}, status=status.HTTP_200_OK)
+
+        return Response(
+            {"success": True, "message": "Search results retrieved successfully.", "data": {"items": items}},
+            status=status.HTTP_200_OK)
 
 
 class FilterItemView(APIView):
@@ -119,4 +182,7 @@ class FilterItemView(APIView):
         items = ItemModel.filter_by_category(category)
         for item in items:
             item["_id"] = str(item["_id"])
-        return Response({"items": items}, status=status.HTTP_200_OK)
+
+        return Response(
+            {"success": True, "message": "Filtered items retrieved successfully.", "data": {"items": items}},
+            status=status.HTTP_200_OK)
